requirements.txt
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
Flask>=2.3.0
